requirements.txt
PyQt5~=5.15.9
python-dateutil~=2.8.2
urllib3~=1.26.4
setuptools~=49.2.1